Add test_sercomlib sketch
This defines some (well "a") macros to make it easier to set up the unallocated SERCOM on (for example) SAMD51-based boards as Serial (UART) interfaces.

Currently, this works best on the Adafruit Grand Central Station Board, which defines almost all of the needed symbols, and only needs a bit of connection.